fix date picker
